+++ src/ConfectioneryPack/Sugar.java
@@ -4,7 +4,7 @@ public class Sugar extends Confection {
     private final SugarType sugarType;
     private int spoons;
 
-    public Sugar(long _number, float _calories, boolean _availability, SugarType sugar, int quantity) {
+    public Sugar(int _number, float _calories, boolean _availability, SugarType sugar, int quantity) {
         super(_number, _calories, _availability);
         System.out.println("Used " + sugar.curIndexValue() + " sugar briquette.");
         sugarType = sugar;
